done with resolving errors

+++ backend/ecommerce/src/main/java/com/example/ecommerce/payload/ProductDTO.java
@@ -24,8 +24,8 @@ public class ProductDTO{
       String description;
   
       Integer quantity;
-      Integer price;
-      Integer discount;
-      Integer specialPrice;
+      double price;
+      double discount;
+      double specialPrice;
   
 }
