+++ backend/ecommerce/src/main/java/com/example/ecommerce/payload/ProductResponse.java
@@ -1,5 +1,11 @@
 package com.example.ecommerce.payload;
 
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -8,7 +14,7 @@ public class ProductResponse{
       List<ProductDTO> content;
       Integer pageNumber;
       Integer pageSize;
-      Integer totalElements;
+      Long totalElements;
       Integer totalPages;
       boolean isLast;
 }
